Break up of ConfigurationManager to improve maintainability. The logic to find the right Configuration class (now that wild cards are supported) is isolated. In addition, the java doc is updated for various classes.
