+++ src/main/java/net/technolords/micro/command/ResetCommand.java
@@ -13,6 +13,13 @@ public class ResetCommand {
     private static final Logger LOGGER = LoggerFactory.getLogger(StatsCommand.class);
     private static StatisticsHandler statisticsHandler;
 
+    /**
+     * Auxiliary method that resets the statistics. Note that the StatisticsHandler is fetched from the Registry, which
+     * is in fact a Jetty component.
+     *
+     * @return
+     *  The result of the reset command.
+     */
     public static ResponseContext executeCommand() {
         LOGGER.debug("Reset command called...");
         ResponseContext responseContext = new ResponseContext();
