+++ src/main/java/net/technolords/micro/command/StatsCommand.java
@@ -13,6 +13,17 @@ public class StatsCommand {
     private static final Logger LOGGER = LoggerFactory.getLogger(StatsCommand.class);
     private static StatisticsHandler statisticsHandler;
 
+    /**
+     * Auxiliary method that reports the statistics. Note that the StatisticsHandler is fetched from the Registry,
+     * which is in fact a Jetty component.
+     *
+     * @param type
+     *  The type associated with the report/stats. When the type is 'html' the report is generated in HTML format,
+     *  otherwise a custom String is returned.
+     *
+     * @return
+     *  The result of the stats command.
+     */
     public static ResponseContext executeCommand(String type) {
         LOGGER.debug("Stats command called...");
         ResponseContext responseContext = new ResponseContext();
