+++ src/main/java/net/technolords/micro/command/StopCommand.java
@@ -11,6 +11,16 @@ import net.technolords.micro.domain.ResponseContext;
 public class StopCommand {
     private static final Logger LOGGER = LoggerFactory.getLogger(StopCommand.class);
 
+    /**
+     * Auxiliary method that stops the Main execution. This is achieved by calling stop on the CamelContext
+     * which is associated with the Main.
+     *
+     * @param exchange
+     *  The exchange associated with the stop command.
+     *
+     * @return
+     *  The result of the stop command (unlikely to be received, as the execution is terminating).
+     */
     public static ResponseContext executeCommand(Exchange exchange) {
         LOGGER.debug("Stop command called...");
         ResponseContext responseContext = new ResponseContext();
