+++ src/main/java/net/technolords/micro/input/xml/ConfigurationToNamespaceContext.java
@@ -1,4 +1,4 @@
-package net.technolords.micro.config;
+package net.technolords.micro.input.xml;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -8,6 +8,7 @@ import javax.xml.namespace.NamespaceContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import net.technolords.micro.config.ConfigurationManager;
 import net.technolords.micro.domain.jaxb.Configuration;
 import net.technolords.micro.domain.jaxb.namespace.NamespaceConfig;
 import net.technolords.micro.domain.jaxb.namespace.NamespaceList;
