+++ src/test/java/net/technolords/micro/command/CommandManagerTest.java
@@ -1,5 +1,5 @@
 package net.technolords.micro.command;
 
 public class CommandManagerTest {
-
+    // TODO: implement (at least the unsupported command part)
 }
